+++ src/hooks/postrun/notify.ts
@@ -1,7 +1,5 @@
 import * as c from "chalk";
-import { elapseEnd, getCurrentGitBranch, getGitRepoName, uxLog } from "../../common/utils";
-import { canSendNotifications, sendNotification } from "../../common/utils/notifUtils";
-import { MetadataUtils } from "../../common/metadata-utils";
+import { elapseEnd, uxLog } from "../../common/utils";
 
 // The use of this method is deprecated: use NotifProvider.sendNotification :)
 
@@ -31,27 +29,5 @@ export const hook = async (options: any) => {
       }
     }
     globalThis.webSocketClient = null;
-  }
-
-  // Send hook to microsoft ?teams if MS_TEAMS_WEBHOOK_URL env var is set, or msTeamsWebhookUrl in config
-  if ((await canSendNotifications()) && options?.Command?.triggerNotification === true) {
-    const diffFiles = await MetadataUtils.listChangedFiles();
-    // No notif if no updated file
-    if (diffFiles.length === 0) {
-      return;
-    }
-    // Send WebHook
-    const jobUrl = process.env.CI_JOB_URL || "Missing CI_JOB_URL variable";
-    const projectName = process.env.CI_PROJECT_NAME || (await getGitRepoName()) || "Missing CI_PROJECT_NAME variable";
-    const branchName = process.env.CI_COMMIT_REF_NAME || (await getCurrentGitBranch({ formatted: true })) || "Missing CI_COMMIT_REF_NAME variable";
-    const envName = projectName + "/" + branchName;
-    const diffFilesTxt = diffFiles.map((diffFile) => `• ${diffFile.path.replace("force-app/main/default/", "")} (${diffFile.index})`).join("\n");
-    await sendNotification({
-      title: `Updates detected in org ${envName}`,
-      text: `<pre>${diffFilesTxt}</pre>`,
-      summary: `Changes on metadatas has been detected on ${envName}. You may want to have a look !`,
-      buttons: [{ title: "View commit", url: jobUrl }],
-      severity: "info",
-    });
   }
 };